Added method ``ImageData<Vec<u32>>::from_buffer()``

+++ tests/readme_examples.rs
@@ -25,7 +25,7 @@ fn resize_lanczos3(src_pixels: &[u8], width: NonZeroU32, height: NonZeroU32) -> 
     dst_image.get_buffer().to_owned()
 }
 
-fn crop_and_resize_image(mut src_view: SrcImageView) -> ImageData<Vec<u8>> {
+fn crop_and_resize_image(mut src_view: SrcImageView) -> ImageData<Vec<u32>> {
     // Set crop-box for view of source image
     src_view
         .set_crop_box(CropBox {
